% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{forecast.crbm}
\alias{forecast.crbm}
\title{Predicting using a Conditional Restricted Boltzmann Machine Function}
\usage{
forecast.crbm(crbm, sequence, n_samples = 1, n_gibbs = 30)
}
\arguments{
\item{crbm.}{A trained CRBM using train.crbm() function.}

\item{sequence.}{Sequence for first input and its history. A matrix with data,
one example per row. Must contain more rows than crbm delay.}

\item{n_samples.}{Number of samples to generate forward. Default = 1.}

\item{n_gibbs.}{Number of alternating Gibbs steps per iteration. Default = 30-}
}
\description{
Function to, given initialization of visibles and matching history, generate
n_samples in future.
}
\examples{
sequence <- dataset[1:(crbm$delay+1), ];
res2 <- forecast.crbm(crbm, sequence, 50, 30);
}
\keyword{CRBM,}
\keyword{FORECAST}
\keyword{RBM,}
