% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{predict.crbm}
\alias{predict.crbm}
\title{Predicting using a Conditional Restricted Boltzmann Machine Function}
\usage{
\method{predict}{crbm}(crbm, newdata)
}
\arguments{
\item{crbm.}{A trained CRBM using train.crbm() function.}

\item{newdata.}{A matrix with data, one example per row. Must contain
more rows than crbm delay.}
}
\description{
Function to Predict data from a CRBM. Returns two matrices: activation
and reconstruction.
}
\examples{
test_X <- t(array(c(1, 1, 0, 0, 0, 0,
                    0, 1, 1, 1, 0, 0,
                    0, 0, 0, 1, 1, 0), c(6,3)));
res <- predict.crbm(crbm1, test_X);
## Also works as
res <- predict(crbm1, test_X);
}
\keyword{CRBM,}
\keyword{RBM}
