% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{train.rbm}
\alias{train.rbm}
\title{Training a Restricted Boltzmann Machine Function}
\usage{
train.rbm(dataset, batch_size = 1, n_hidden = 3, training_epochs = 1000,
  learning_rate = 0.01, momentum = 0.8, rand_seed = 1234)
}
\arguments{
\item{dataset.}{A matrix with data, one example per row.}

\item{batch_size.}{Number of examples per training mini-batch. Default = 1.}

\item{n_hidden.}{Number of hidden units in the RBM. Default = 3.}

\item{training_epochs.}{Number of training epochs. Default = 1000.}

\item{learning_rate.}{The learning rate for training. Default = 0.01.}

\item{momentum.}{The momentum for training. Default = 0.8.}

\item{rand_seed.}{Random seed. Default = 1234.}
}
\description{
This function trains a RBM. Returns a RBM in list form
}
\examples{
train_X <- t(array(c(1, 1, 1, 0, 0, 0,
                     1, 0, 1, 0, 0, 0,
                     1, 1, 1, 0, 0, 0,
                     0, 0, 1, 1, 1, 0,
                     0, 0, 1, 0, 1, 0,
                     0, 0, 1, 1, 1, 0), c(6, 6)));
rbm1 <- train.rbm(train_X);
}
\keyword{RBM}
